query_me/src/requirements.txt
flask
flask-cors
transformers
torch
mysql-connector-python
tiktoken>=0.5.1
